Move grafana upgrade
