+++ upgrades/0.0.78.bump-grafana/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/oslokommune/okctl-upgrade/0.0.78.bump-grafana/pkg/commonerrors"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.78.bump-grafana/pkg/commonerrors"
 	"github.com/spf13/cobra"
 )
