+++ upgrades/0.0.78.bump-grafana/pkg/grafana/api.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/oslokommune/okctl-upgrade/0.0.78.bump-grafana/pkg/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.78.bump-grafana/pkg/logger"
 )
 
 // Upgrader is a sample okctl component
