+++ upgrades/0.0.78.bump-grafana/pkg/grafana/kubectl.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/oslokommune/okctl-upgrade/0.0.78.bump-grafana/pkg/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.78.bump-grafana/pkg/logger"
 
 	"github.com/Masterminds/semver"
